packages/bridge/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
inquirer
